MAYBE REVERT ??? this may be pointless ???
jumped through some hoops to replicate the source so that alias gets
source for new and old identifier

maybe pointless here since they are the same

OTOH, tools like Swish Lint may want to be able to determine what the
alias means at the source level and this might sort of help untangle?

Hmm. Can we find a way to get the source of the actual *internal* binding?