WIP try to fix make -j issue
Original symptom:

  This wasn't working at top-level of repo since around 8777b4f7ac
  $ ./configure; make clean; make -j safe-check

I think it works now, but there are lots of cases to test:
  - change Scheme version via ./configure
  - keep Scheme configuration but change hash for that Scheme
  - modify Swish source file
  - change Swish version by making another commit (e.g., gmark)
  - make again after no changes
  - make safe-check
  - make coverage
  - make test
  - make -C src/swish mat-prereq
  - cd src; ./go

We'd want to try all of those (except maybe ./go) with make -j and make -j0

I left in some verbose debugging stuff echoed to console so we could see
the machinery at work. One mechanism here is checking MAKE_RESTARTS to
figure out whether or not we should be removing the .tmp-include files.

We need to remove them, but the notion here is to remove them only if we're
the initial call to make, not some sub $(MAKE) call.

I had hoped that .INTERMEDIATE: would handle all of that transparently, but
then the .tmp-include files simply don't have the effect we want.

Another notion we might try is to use MAKE_RESTARTS to figure out if we're
the initial call to make and have that call add some explicit dependency that
does the version checking.